robot-env/requirements.txt
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
RPi.GPIO==0.7.1
RPLCD==1.4.0
smbus2==0.5.0
